# Conflicts:
#	ComGraphics/ComGraphics/ComGraphics.vcxproj.filters

+++ ComGraphics/ComGraphics/Source/MobGhost.cpp
@@ -107,5 +107,5 @@ void MobGhost::setSpawnGhost(float xpos, float zpos)
 void MobGhost::TakeDamage(int damage)
 {
     std::cout << "hit" << std::endl;
-	Explorer::instance()->MinusHP(damage);  
+	Explorer::instance()->MinusHP((float)damage);  
 }
